Questions added describing fork, clone and branches
Fork allows you to make a copy of someone else repository directly on GitHub servers without pulling down the code to your local machine first while Clone is a copy of your repository. Like the example of the recipe	repository, if i want to modify the recipe without tempering with the original code, i will fork the recipe repository and clone it directly to my local machine.